Added lots of new RIF built-ins. Updated hierarchy of numeric terms in order to be compatible with XSD. All numeric terms now use BigDecimal in order to allow arbitrary-precision values for specific data types.

+++ iris/trunk/src/org/deri/iris/builtins/string/StringSubstringBuiltin.java
@@ -63,7 +63,8 @@ public class StringSubstringBuiltin extends FunctionalBuiltin {
 			IIntegerTerm endIndex = (IIntegerTerm) terms[2];
 
 			String substring = string.getValue().substring(
-					beginIndex.getValue(), endIndex.getValue());
+					beginIndex.getValue().intValue(),
+					endIndex.getValue().intValue());
 
 			return Factory.TERM.createString(substring);
 		}
